- be less intrusive with sck/DataStream.h
