+++ dbus/DBusMessage.cc
@@ -433,28 +433,4 @@ namespace DBus
 	return hoho;
     }
 
-    Hihi&
-    operator>>(Hihi& hihi, sck::SocketFd& data)
-    {
-	if (hihi.get_type() != DBUS_TYPE_UNIX_FD)
-	    throw MarshallingException();
-
-	int fd;
-	dbus_message_iter_get_basic(hihi.top(), &fd);
-	dbus_message_iter_next(hihi.top());
-	data.set_fd(fd);
-
-	return hihi;
-    }
-
-    Hoho&
-    operator<<(Hoho& hoho, const sck::SocketFd& data)
-    {
-	const int fd = data.get_fd();
-	if (!dbus_message_iter_append_basic(hoho.top(), DBUS_TYPE_UNIX_FD, &fd))
-	    throw FatalException();
-
-	return hoho;
-    }
-
 }
