+++ dbus/DBusMessage.h
@@ -31,7 +31,6 @@
 #include <list>
 #include <map>
 
-#include "sck/DataStream.h"
 
 namespace DBus
 {
@@ -243,8 +242,6 @@ namespace DBus
     Hihi& operator>>(Hihi& hihi, map<string, string>& data);
     Hoho& operator<<(Hoho& hoho, const map<string, string>& data);
 
-    Hihi& operator>>(Hihi& hihi, sck::SocketFd& data);
-    Hoho& operator<<(Hoho& hoho, const sck::SocketFd& data);
 
     template <typename Type>
     Hihi& operator>>(Hihi& hihi, vector<Type>& data)
