+++ sck/Makefile.am
@@ -7,6 +7,6 @@ AM_CPPFLAGS = -I$(top_srcdir)
 noinst_LTLIBRARIES = libsck.la
 
 libsck_la_SOURCES =					\
-	DataStream.cc		DataStream.h
+	Socket.cc		Socket.h
 
-libsck_la_LIBADD = -lboost_system -lboost_serialization
+libsck_la_LIBADD = -lboost_system
