+++ server/Types.cc
@@ -101,4 +101,14 @@ namespace DBus
 	return hoho;
     }
 
+    Hoho&
+    operator<<(Hoho& hoho, const SocketFd& data)
+    {
+	const int fd = data.get_fd();
+	if (!dbus_message_iter_append_basic(hoho.top(), DBUS_TYPE_UNIX_FD, &fd))
+	    throw FatalException();
+
+	return hoho;
+    }
+
 }
